Extract simulator into separate bin module
